client: return int from LRUDict.time_fn
In python3, round returns int while it returns float in python2.
Let me change the function always return int.

Also use int in the test to prevent unwanted type change to float.

Bug: 1010816
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-py/+/2272196
Commit-Queue: Junji Watanabe <jwata@google.com>
Auto-Submit: Takuto Ikuta <tikuta@chromium.org>

+++ client/utils/lru.py
@@ -28,7 +28,7 @@ class LRUDict(object):
 
     Can be substituted in individual LRUDict instances, especially for tests.
     """
-    return round(time.time())
+    return int(round(time.time()))
 
   def __init__(self):
     # Ordered key -> (value, timestamp) mapping,
